libc: Allow a persist property to modify the value of verifiedbootstate
If persist.hide_verifiedbs is set to 1 then ro.boot.verifiedbootstate
will be returned as "green" no matter what it is really set to.

This would allow Settings to have a preference to toggle whether or not
to hide the verifiedbootstate to temporarily pass safetynet.

Taken from https://gerrit.aospa.co/#/c/3396/